chore(libvmaf): CodeQL C bulk sweep — 60/78 deferred alerts fixed (#552)
Resolves the 60 mechanically-fixable alerts deferred from #538 across three
CodeQL categories on upstream-mirror C files:

* cpp/integer-multiplication-cast-to-long (44/44):
  prefixed widening cast on the LHS operand before multiplication —
  (size_t), (uint64_t), (double), (long), (ptrdiff_t) — never wrapped
  around the whole expression. Touched y4m_input.c (13), cambi.c (5),
  integer_psnr.c (4), iqa/convolve.c (deferred — bit-exactness gate),
  vmaf_bench.c (4), psnr_hvs.c (2), ansnr_avx2.c (2), ansnr_avx512.c (2),
  ciede.c (2), pdjson.c (1), yuv_input.c (1), adm_tools.h (1),
  integer_adm.c (1), adm_avx2.c (1), adm_avx512.c (1).
* cpp/commented-out-code (19/19): deleted dead/disabled code blocks
  (TODO-with-corpse-code, debug-print scaffolding, AVX-512 VP-loop dead
  variant, libsvm dead alternate). Per CLAUDE no-corpse-comments rule.
* cpp/large-parameter (1/15): converted cambi_score(... CambiBuffers
  buffers, ...) → const CambiBuffers *. The 14 remaining alerts on the
  VifBuffer SIMD ABI (integer_vif.{c,h}, x86/vif_avx2.c, x86/vif_avx512.c)
  are deferred to a coordinated multi-backend follow-up — the calling
  convention is shared with arm64/vif_neon.c (not in this sweep) and
  flipping it requires lockstep updates across NEON + AVX2 + AVX-512 +
  scalar paths.

Touched-file lint-clean per CLAUDE §12 r12. One unrelated IWYU warning
absorbed: dropped unused config.h include in libvmaf/test/test_cpu.c
after deleting the AVX2 mask-test corpse that gated it. Build + 50 C
unit tests pass on CPU configuration. Netflix golden gate runs identical
9 pre-existing failures as master (PSNR/NIQE Python harness, unrelated).

Coordinates with parallel sibling sweep on cpp/declaration-hides-variable
(64 alerts, branch agent-a5fe363b26a4d7592) — touches many of the same
files; whichever lands first, the other rebases.

+++ libvmaf/src/feature/adm_tools.h
@@ -347,7 +347,7 @@ static FORCE_INLINE float dwt_quant_step(const struct dwt_model_params *params, 
 
     // Formula (9), page 1171
     float temp = log10(pow(2.0, lambda + 1) * params->f0 * params->g[theta] / r);
-    float Q = 2.0 * params->a * pow(10.0, params->k * temp * temp) /
+    float Q = 2.0 * params->a * pow(10.0, params->k * (double)temp * temp) /
               dwt_7_9_basis_function_amplitudes[lambda][theta];
 
     return Q;
